+++ docs/api/admin/code_samples/JavaScript/admin_inventory-items/post.js
@@ -2,7 +2,7 @@ import Medusa from "@medusajs/medusa-js"
 const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
 // must be previously logged in or use api token
 medusa.admin.inventoryItems.create({
-  variant_id: 'variant_123',
+  variant_id: "variant_123",
 })
 .then(({ inventory_item }) => {
   console.log(inventory_item.id);