+++ docs/api/admin/code_samples/JavaScript/admin_regions_{id}/post.js
@@ -1,8 +1,8 @@
 import Medusa from "@medusajs/medusa-js"
 const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
 // must be previously logged in or use api token
-medusa.admin.regions.update(region_id, {
-  name: 'Europe'
+medusa.admin.regions.update(regionId, {
+  name: "Europe"
 })
 .then(({ region }) => {
   console.log(region.id);