+++ docs/api/admin/code_samples/Shell/admin_inventory-items_{id}/delete.sh
@@ -1,2 +1,2 @@
-curl --location --request DELETE 'https://medusa-url.com/admin/inventory-items/{id}' \
---header 'Authorization: Bearer {api_token}'
+curl -X DELETE 'https://medusa-url.com/admin/inventory-items/{id}' \
+-H 'Authorization: Bearer {api_token}'