+++ docs/api/admin/code_samples/Shell/admin_orders_{id}_claims_{claim_id}_fulfillments/post.sh
@@ -1,2 +1,2 @@
-curl --location --request POST 'https://medusa-url.com/admin/orders/{id}/claims/{claim_id}/fulfillments' \
---header 'Authorization: Bearer {api_token}'
+curl -X POST 'https://medusa-url.com/admin/orders/{id}/claims/{claim_id}/fulfillments' \
+-H 'Authorization: Bearer {api_token}'