+++ docs/api/admin/code_samples/Shell/admin_products_{id}_options_{option_id}/delete.sh
@@ -1,2 +1,2 @@
-curl --location --request DELETE 'https://medusa-url.com/admin/products/{id}/options/{option_id}' \
---header 'Authorization: Bearer {api_token}'
+curl -X DELETE 'https://medusa-url.com/admin/products/{id}/options/{option_id}' \
+-H 'Authorization: Bearer {api_token}'