+++ docs/api/admin/code_samples/Shell/admin_publishable-api-keys_{id}_sales-channels_batch/delete.sh
@@ -1,6 +1,6 @@
-curl --location --request DELETE 'https://medusa-url.com/admin/publishable-api-keys/{pka_id}/batch' \
---header 'Authorization: Bearer {api_token}' \
---header 'Content-Type: application/json' \
+curl -X DELETE 'https://medusa-url.com/admin/publishable-api-keys/{id}/batch' \
+-H 'Authorization: Bearer {api_token}' \
+-H 'Content-Type: application/json' \
 --data-raw '{
     "sales_channel_ids": [
       {