+++ docs/api/admin/code_samples/Shell/admin_stock-locations/get.sh
@@ -1,2 +1,2 @@
-curl --location --request GET 'https://medusa-url.com/admin/stock-locations' \
---header 'Authorization: Bearer {api_token}'
+curl 'https://medusa-url.com/admin/stock-locations' \
+-H 'Authorization: Bearer {api_token}'