+++ docs/api/store/code_samples/Shell/store_auth_{email}/get.sh
@@ -1,2 +1,1 @@
-curl --location --request GET 'https://medusa-url.com/store/auth/user@example.com' \
---header 'Cookie: connect.sid={sid}'
+curl 'https://medusa-url.com/store/auth/user@example.com'