+++ docs/api/store/code_samples/Shell/store_carts_{id}_discounts_{code}/delete.sh
@@ -1,1 +1,1 @@
-curl --location --request DELETE 'https://medusa-url.com/store/carts/{id}/discounts/{code}'
+curl -X DELETE 'https://medusa-url.com/store/carts/{id}/discounts/{code}'