+++ docs/api/store/code_samples/Shell/store_customers_me_orders/get.sh
@@ -1,2 +1,2 @@
-curl --location --request GET 'https://medusa-url.com/store/customers/me/orders' \
---header 'Cookie: connect.sid={sid}'
+curl 'https://medusa-url.com/store/customers/me/orders' \
+-H 'Cookie: connect.sid={sid}'