+++ packages/workflows/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @medusajs/workflows
 
+## 0.1.1
+
+### Patch Changes
+
+- [#4673](https://github.com/medusajs/medusa/pull/4673) [`e78c47b66`](https://github.com/medusajs/medusa/commit/e78c47b66fe99f5d242ca119dea580804a31768c) Thanks [@adrien2p](https://github.com/adrien2p)! - Feat/workflow confusion elimination
+
+- [#4689](https://github.com/medusajs/medusa/pull/4689) [`43f34866c`](https://github.com/medusajs/medusa/commit/43f34866c817da43ca2b865fd58e5fc8c3dc891b) Thanks [@adrien2p](https://github.com/adrien2p)! - feat(workflows): Improve typings
+
+- Updated dependencies [[`4b80ba8a3`](https://github.com/medusajs/medusa/commit/4b80ba8a356806da3c92634e40e8946da25e35ee), [`8af55aed8`](https://github.com/medusajs/medusa/commit/8af55aed87da7252c7c261175bc98331466a0da8), [`5c60aad17`](https://github.com/medusajs/medusa/commit/5c60aad177a99574ffff5ebdc02ce9dc86ef9af9), [`4073b7313`](https://github.com/medusajs/medusa/commit/4073b73130c874dc7d2240726224a01b7b19b1a1)]:
+  - @medusajs/modules-sdk@1.9.1
+  - @medusajs/utils@1.9.5
+
 ## 0.1.0
 
 ### Minor Changes